Clean up

+++ src/local/clean.sh
@@ -53,18 +53,20 @@ fi
 chillbox_data_home="${XDG_DATA_HOME:-"$HOME/.local/share"}/chillbox/$CHILLBOX_INSTANCE/$WORKSPACE"
 encrypted_secrets_dir="${ENCRYPTED_SECRETS_DIR:-${chillbox_data_home}/encrypted-secrets}"
 
-encrypted_secrets_file_list="$(find "$encrypted_secrets_dir" -type f)"
-if [ -z "$encrypted_secrets_file_list" ]; then
-  printf '\n%s\n' "No encrypted secrets found to delete in chillbox instance '$CHILLBOX_INSTANCE' and workspace '$WORKSPACE'."
-else
-  printf '\n%s\n' "The $0 script will delete the encrypted secrets in the directory '$encrypted_secrets_dir' for the chillbox instance '$CHILLBOX_INSTANCE' and workspace '$WORKSPACE'."
-  printf '\n%s\n' "$encrypted_secrets_file_list"
-  printf '\n%s\n' "Delete the encrypted secrets in the $encrypted_secrets_dir directory? [y/n]"
-  read -r confirm
-  if [ "$confirm" = "y" ]; then
-    find "$encrypted_secrets_dir" -type f -delete
+if [ -d "$encrypted_secrets_dir" ]; then
+  encrypted_secrets_file_list="$(find "$encrypted_secrets_dir" -type f)"
+  if [ -z "$encrypted_secrets_file_list" ]; then
+    printf '\n%s\n' "No encrypted secrets found to delete in chillbox instance '$CHILLBOX_INSTANCE' and workspace '$WORKSPACE'."
   else
-    printf '\n%s\n' "Skipping deletion of encrypted secrets in $encrypted_secrets_dir directory."
+    printf '\n%s\n' "The $0 script will delete the encrypted secrets in the directory '$encrypted_secrets_dir' for the chillbox instance '$CHILLBOX_INSTANCE' and workspace '$WORKSPACE'."
+    printf '\n%s\n' "$encrypted_secrets_file_list"
+    printf '\n%s\n' "Delete the encrypted secrets in the $encrypted_secrets_dir directory? [y/n]"
+    read -r confirm
+    if [ "$confirm" = "y" ]; then
+      find "$encrypted_secrets_dir" -type f -delete
+    else
+      printf '\n%s\n' "Skipping deletion of encrypted secrets in $encrypted_secrets_dir directory."
+    fi
   fi
 fi
